feat(metanode): Enable rollback from 3.4.0 to 3.3.2
1) Add protocal opFSMSentToChanWithVer, enable compatible during update version because []proto.ExtentKey has no version info
2) Update inode marshal and unmarshal process, do not persist version 3 flag if not eable snapshot

+++ metanode/partition_free_list.go
@@ -554,6 +554,11 @@ func (mp *metaPartition) deleteObjExtents(oeks []proto.ObjExtentKey) (err error)
 
 	for i := 0; i < total; i += maxDelCntOnce {
 		max := util.Min(i+maxDelCntOnce, total)
+		if mp.ebsClient == nil {
+			err = fmt.Errorf("(%v) error ebsClient nil", mp.config.PartitionId)
+			log.LogErrorf("deleteObjExtents %v", err)
+			return
+		}
 		err = mp.ebsClient.Delete(oeks[i:max])
 		if err != nil {
 			log.LogErrorf("[deleteObjExtents] vol(%v) mp(%v) delete ebs eks fail, cnt(%d), err(%s)", mp.config.VolName, mp.config.PartitionId, max-i, err.Error())
